Floor building feature of class Floor factored out into its own class, InMemoryFloorBuilder, which implements interface FloorBuilder.
