+++ sensor/src/main/java/com/team4/sensor/FloorBuilder.java
@@ -1,0 +1,5 @@
+package com.team4.sensor;
+
+public interface FloorBuilder {
+    void buildFloor();
+}
